Devyanshu_project/requirements.txt
Flask==2.0.2
openpyxl==3.0.9
pandas==1.3.3
PyPDF2==1.26.0
python-docx==0.8.11
